Update CreateTablePlan to include dataDir parameter for flexible file path handling
This commit modifies the CreateTablePlan struct to add a dataDir field, allowing for dynamic specification of the data directory when creating heap files. The NewCreateTablePlan constructor is updated to accept this new parameter, enhancing the flexibility of file path management in the createHeapFile method. These changes improve the usability of the CreateTablePlan in various contexts.